Update `context.json` to allow for typing.
## What? 
Make context.json a templated function, allowing for better Typing support on route responses, if the `.json` function is used.

## Why?
To be able to have strongly typed responses with a single function call.


    interface ResponseInterface {
        hello: string;
    }
    
    interface ErrorInterface {
        success: boolean;
        error: string;
    }
    
    const get = async (ctx: Context): Promise<ResponseInterface | ErrorInterface> => {
        if (1 !== 0) {
                return ctx.json({success: false, error: 'Not Found'}, Status.NotFound);
        }
        return ctx.json({hello: 'World'});
    };

Considering this code block, and the void return of context.json now, this would result in a linting error when considering the promise return type. 

## Implementation Example

    interface ResponseInterface {
        hello: string;
    }
    
    interface ErrorInterface {
        success: boolean;
        error: string;
    }
    
    const get = async (ctx: Context): Promise<ResponseInterface | ErrorInterface> => {
        if (1 !== 0) {
                return ctx.json<ErrorInterface>({success: false, error: 'Not Found'}, Status.NotFound);
        }
        return ctx.json<ResponseInterface>({hello: 'World'});
    };


## How does the proposal fix the why?
Making context.json a templated function will allow us to return from that function, our intended output data ('v'), withe the type (T) that we expect, this allows us to `return context.json` to 'early exit' from various code where required, and have strongly typed responses expectations from controllers. 

Yes, we can set the code, and return a JSON object from an async function already, however using the function handles this all in one go, and is more intuitive/meaningful (as a function)

## Alternatives
Currently to get this functionality into your ABC application, you will need a wrapper. Or a helper function that mimics this flow, through a CustomContext or middleware or other means that embeds a context reference, exposing `context.json` 

    function  json<T>(v: T, code: Status = Status.OK): T {
	    // context.json(v,code);
	    return  v;
    }